+++ web/src/main.cpp
@@ -46,8 +46,12 @@ int main(int argc, char** argv) {
             return 1;
         }
 
-        // Init ranker
-        mithril::ranking::InitRanker(index_path);
+        try {
+            // Init ranker
+            mithril::ranking::InitRanker(index_path);
+        } catch (std::exception& e){
+            spdlog::info("ranker was not initialized, make sure this is distributed one");
+        }
 
         // Create and init search plugin with both paths
         auto search_plugin = new SearchPlugin(server_config_path, index_path);