+++ CHANGELOG.md
@@ -132,6 +132,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Strip all HTML from title on frontend (#168)
 - Fix hardcoded path (#171)
 - optimize usage of positions in ranker (#170)
+- downstream fixes after ranking (#172)
 
 ## [0.1.0] - 2025-02-07